backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0

python-dotenv==1.0.1

backboard-sdk==1.3.3  # Backboard Python SDK [web:43]
httpx==0.27.0

pydantic==2.9.0 # Clean request/response model
